modify package names
