+++ src/main/java/com/team1/dealerApp/user/CreateUserDTO.java
@@ -1,8 +1,7 @@
-package com.team1.dealerApp.models.dtos;
+package com.team1.dealerApp.user;
 
-import com.team1.dealerApp.entities.Movie;
-import com.team1.dealerApp.entities.TvShow;
-import com.team1.dealerApp.models.SubsciprionStatus;
+import com.team1.dealerApp.video.movie.Movie;
+import com.team1.dealerApp.video.tvshow.TvShow;
 import lombok.*;
 
 import java.util.List;
