+++ src/main/java/com/team1/dealerApp/user/UserRepository.java
@@ -1,6 +1,5 @@
-package com.team1.dealerApp.repositories;
+package com.team1.dealerApp.user;
 
-import com.team1.dealerApp.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
