+++ src/main/java/com/team1/dealerApp/video/Genre.java
@@ -1,4 +1,4 @@
-package com.team1.dealerApp.models;
+package com.team1.dealerApp.video;
 
 public enum Genre {
     ACTION,
