+++ src/main/java/com/team1/dealerApp/video/VideoStatus.java
@@ -1,4 +1,4 @@
-package com.team1.dealerApp.models;
+package com.team1.dealerApp.video;
 
 public enum VideoStatus {
     NOT_AVAILABLE,
