+++ src/main/java/com/team1/dealerApp/video/tvshow/TvShowDTO.java
@@ -1,8 +1,7 @@
-package com.team1.dealerApp.models.dtos;
+package com.team1.dealerApp.video.tvshow;
 
-import com.team1.dealerApp.models.Genre;
-import com.team1.dealerApp.models.VideoStatus;
-
+import com.team1.dealerApp.video.Genre;
+import com.team1.dealerApp.video.VideoStatus;
 import lombok.*;
 
 import java.time.Year;
@@ -10,14 +9,12 @@ import java.util.List;
 
 @Setter
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @Builder
-public class MovieDTO {
+public class TvShowDTO {
 
     private String title;
-
-    private int runningTime;
 
     private Genre genre;
 
@@ -25,7 +22,7 @@ public class MovieDTO {
 
     private String director;
 
-    private Year year;
+    private Year releaseYear;
 
     private double purchasePrice;
 
@@ -34,6 +31,10 @@ public class MovieDTO {
     private String plot;
 
     private float rating;
+
+    private int season;
+
+    private int episode;
 
     private VideoStatus videoStatus;
