+++ src/main/java/com/team1/dealerApp/video/tvshow/TvShowRepository.java
@@ -1,6 +1,5 @@
-package com.team1.dealerApp.repositories;
+package com.team1.dealerApp.video.tvshow;
 
-import com.team1.dealerApp.entities.TvShow;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TvShowRepository extends JpaRepository<TvShow, Long> {
